@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'rails', '6.0.4.1'
+gem 'rails', '7.0.0'
 gem 'rake', '13.0.1'
 
 group :development do
@@ -15,39 +15,39 @@ end
 group :development, :test do
   gem 'byebug', '~> 11.1.3'
   gem 'dotenv-rails', '~> 2.7.6'
-  gem 'factory_bot_rails', '~> 6.1.0'
-  gem 'rspec-rails', '~> 4.0.1'
+  gem 'factory_bot_rails', '~> 6.2.0'
+  gem 'rspec-rails', '~> 5.0.0'
 end
 
 group :test do
-  gem 'pundit-matchers', '~> 1.6.0'
+  gem 'pundit-matchers', '~> 1.7.0'
   gem 'rspec_junit_formatter' # Used for formatting CircleCI output
-  gem 'shoulda-matchers', '~> 4.4.1'
+  gem 'shoulda-matchers', '~> 5.1.0'
 end
 
-gem 'bcrypt', '~> 3.1.15'
-gem 'bootsnap', '~> 1.4.8'
-gem 'bootstrap', '~> 4.5.2'
-gem 'cocoon', '~> 1.2.14'
-gem 'devise', '~> 4.7.2'
-gem 'font-awesome-sass', '~> 5.13.0'
+gem 'bcrypt', '~> 3.1.16'
+gem 'bootsnap', '~> 1.9.3'
+gem 'bootstrap', '~> 4.6.1'
+gem 'cocoon', '~> 1.2.15'
+gem 'devise', '~> 4.8.1'
+gem 'font-awesome-sass', '~> 5.15.1'
 gem 'jquery-rails', '~> 4.4.0'
-gem 'kaminari', '~> 1.2.1'
+gem 'kaminari', '~> 1.2.2'
 gem 'kramdown', '~> 2.3.1'
 gem 'local_time', '~> 2.1.0'
-gem 'mini_magick', '~> 4.10.1'
+gem 'mini_magick', '~> 4.11.0'
 gem 'mini_racer', '~> 0.3.1'
 gem 'nokogiri', '1.12.5'
 gem 'onebox', '~> 2.2.19'
-gem 'kt-paperclip', '~> 6.3.0'
-gem 'paper_trail', '~> 11.0.0'
+gem 'kt-paperclip', git: 'https://github.com/kreeti/kt-paperclip', branch: 'master'
+gem 'paper_trail', '~> 12.1.0'
 gem 'pg', '~> 1.2.3'
-gem 'pundit', '~> 2.1.0'
-gem 'puma', '~> 5.5.1'
+gem 'pundit', '~> 2.1.1'
+gem 'puma', '~> 5.5.2'
 gem 'rails_autolink', '~> 1.1.6'
-gem 'ransack', '~> 2.3.2'
-gem 'sanitize', '~> 5.2.1'
-gem 'sassc', '~> 2.3.0'
+gem 'ransack', '~> 2.5.0'
+gem 'sanitize', '~> 6.0.0'
+gem 'sassc', '~> 2.4.0'
 gem 'sassc-rails', '~> 2.1.2'
 gem 'stringex', '~> 2.8.5'
 gem 'uglifier', '>= 4.2.0'
