@@ -38,7 +38,7 @@ class Admin::CategoriesController < ApplicationController
   end
 
   def update
-    if @category.update_attributes category_params
+    if @category.update category_params
       flash[:notice] = 'Successfully updated category.'
       redirect_to_category
     else
